Rename index.jsx files to match naming pattern
[MAILPOET-2450]
