+++ assets/js/src/form_editor/form_editor.jsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
 import Editor from './components/editor.jsx';
-import initStore from './store/index.jsx';
+import initStore from './store/store.jsx';
 
 const appElement = document.querySelector('#mailpoet_form_edit');
